Split the crate
