+++ Cargo.toml
@@ -1,14 +1,11 @@
-[package]
-name = "hello-wasm"
-version = "0.1.0"
-edition = "2021"
-
-[lib]
-crate-type = ["cdylib"]
-
-[dependencies]
-rust-ini = "0.21"
-yaml-rust2 = "0.10"
+[workspace]
+members = [
+    "nix-wasm-rust",
+    "nix-wasm-plugin-fib",
+    "nix-wasm-plugin-ini",
+    "nix-wasm-plugin-yaml"
+]
+resolver = "2"
 
 [profile.release]
 lto = true
