+++ flake.nix
@@ -8,7 +8,7 @@
 
   outputs = { self, ... }@inputs:
     let
-      cargoToml = builtins.fromTOML (builtins.readFile ./Cargo.toml);
+      cargoToml = builtins.fromTOML (builtins.readFile ./nix-wasm-rust/Cargo.toml);
       supportedSystems = [ "aarch64-darwin" "aarch64-linux" "x86_64-darwin" "x86_64-linux" ];
       forAllSystems = f: inputs.nixpkgs.lib.genAttrs supportedSystems (system: f {
         pkgs = import inputs.nixpkgs { inherit system; };
@@ -17,29 +17,46 @@
     in
     {
       packages = forAllSystems ({ pkgs, system }: rec {
-        default = hello-wasm;
-        hello-wasm = with pkgs; rustPlatform.buildRustPackage {
+        default = nix-wasm-plugins;
+
+        nix-wasm-plugins = with pkgs; rustPlatform.buildRustPackage {
           pname = cargoToml.package.name;
           version = cargoToml.package.version;
+
           cargoLock.lockFile = ./Cargo.lock;
+
           src = self;
-          #cargoBuildFlags = "--target wasm32-unknown-unknown";
+
           CARGO_BUILD_TARGET = "wasm32-unknown-unknown";
           buildPhase = "cargo build --release";
+
           checkPhase = "true";
-          installPhase = "cp target/wasm32-unknown-unknown/release/hello_wasm.wasm $out";
+
+          installPhase = ''
+            mkdir -p $out
+            for i in target/wasm32-unknown-unknown/release/*.wasm; do
+              wasm-opt -O3 -o "$out/$(basename "$i")" "$i"
+            done
+          '';
+
           nativeBuildInputs = [
             rustc.llvmPackages.lld
             wasm-bindgen-cli
             wasm-pack
+            binaryen
+          ];
+        };
+      });
 
-            # Dev tools
+      devShells = forAllSystems ({ pkgs, system }: rec {
+        default = with pkgs; self.packages.${system}.default.overrideAttrs (attrs: {
+          nativeBuildInputs = attrs.nativeBuildInputs ++ [
             wabt
             rust-analyzer
             rustfmt
             clippy
           ];
-        };
+        });
       });
     };
 }
