+++ nix-wasm-plugin-fib/Cargo.toml
@@ -1,0 +1,10 @@
+[package]
+name = "nix-wasm-plugin-fib"
+version = "0.1.0"
+edition = "2021"
+
+[lib]
+crate-type = ["cdylib"]
+
+[dependencies]
+nix-wasm-rust = { path = "../nix-wasm-rust" }
