+++ nix-wasm-plugin-ini/Cargo.toml
@@ -1,0 +1,11 @@
+[package]
+name = "nix-wasm-plugin-ini"
+version = "0.1.0"
+edition = "2021"
+
+[lib]
+crate-type = ["cdylib"]
+
+[dependencies]
+nix-wasm-rust = { path = "../nix-wasm-rust" }
+rust-ini = "0.21"
