+++ nix-wasm-plugin-yaml/Cargo.toml
@@ -1,0 +1,11 @@
+[package]
+name = "nix-wasm-plugin-yaml"
+version = "0.1.0"
+edition = "2021"
+
+[lib]
+crate-type = ["cdylib"]
+
+[dependencies]
+nix-wasm-rust = { path = "../nix-wasm-rust" }
+yaml-rust2 = "0.10"
