+++ nix-wasm-rust/Cargo.toml
@@ -1,0 +1,4 @@
+[package]
+name = "nix-wasm-rust"
+version = "0.1.0"
+edition = "2021"
