+++ nix-wasm-rust/src/lib.rs
